refactor(routes): keep general routes colocated
